// src/app/models/budget.d.ts
export interface budgetModel {
  salary: number;
  otherIncome: number;
  taxAmount: number;
  taxPercentage: number;
  retirement: number;
  unemployment: number;
  medicalAid: number;
  insurance: number;
  savings: number;
  investments: number;
  interest: number;
  emergency: number;
  rent: number;
  electricity: number;
  water: number;
  groceries: number;
  fuel: number;
  pet: number;
  bigPurchases: number;
  carEmergencies: number;
  glasses: number;
  clothes: number;
  entertainment: number;
  subscriptions: number;
}